The great package sources refactor of 2015.
The old system of apt-sources was getitng unwieldy and a tad ridiculous, in
  addition to being named poorly (yum sources should go in the same file and
  just use an OS switch to fix things). So, I fixed it.

There is now a core module. This is the one module you cannot remove from your
  top.sls file. It provides common, required state/functionality, in addition
  to containing some reusable states that serve as sorts of mini-functions
  (like core.pass). Every submodule in here is configured entirely through
  pillars and grains and will have an entry file, like pkg-source's
  config-by-distro. The core module's init.sls state includes all major states
  the system needs.

Currently, the only submodule in core is pkg-sources, which is the replacement
  to apt-sources. It configures one generic set of package sources for you
  based on your distro, and then holds all other states that work with package
  sources in core.pkg-sources.sources. Now the files are hidden and still
  quite includable and usable. Other apps located throughout desktop.sls still
  include their own package source state and can require that, the file now is
  just a bit further abstracted into core.

The entry file for pkg-sources, config-by-distro, works by merely determining
  the OS and release of the machine, and requiring a pre-defined package source
  state in core.pkg-sources.distros that sets you up with your basics (e.g., in
  Debian, we turn on non-free; in Debian testing, we enable jessie for you, for
  the purpose of backports).

New pre-defined package source state for Debian testing.

New package source states for debian jessie, wheezy, and sid. New package
  source state for Debian non-free (it is generic and determines your codename
  at run time).

Moved all app-specific source states, like google-chrome and dropbox, to
  core.pkg-sources.sources. Made all of them use a file and not pollute
  sources.list.

Updated all files that referenced apt-sources to point to core.pkg-sources.

New core.pass state that is used for syntatical correctness with include
  statements, see file for more info.

Going to tag this as 0.1.0 for now.